clean code 1_80
# message auto-generated for no-merge-commit merge:
merge clean_code_1_80 into master

Clean Code 1到80条

Created-by: msprofDeveloper

Match-id-4d257c0977b5267882de3ab3f00d589a8f2f5561

+++ collector/dvvp/message/prof_json_config.h
@@ -82,7 +82,7 @@ struct ProfGeOptionsConfig {
         if ((json).contains(inputSwitch)) {                                                                 \
             (json).at(inputSwitch).get_to(intValueFromJson);                                                \
             if (!(intValueFromJson).empty() && Utils::CheckStringIsValidNatureNum(intValueFromJson)) {      \
-                (dataStruct)->field = std::stoi(intValueFromJson);                                          \
+                (dataStruct)->field = static_cast<uint32_t>(std::stoi(intValueFromJson));                   \
             }                                                                                               \
         }                                                                                                   \
     } while (0)
